website: fix docHandler test when run via devcam
The test added for proper markdown title generation
do not work when run via devcam.  This is caused by
depending on doc/uses.md as the input testdata.
devcam does not copy markdown files into the target tmp
directory so this test fails when run that way.

Solution is create a markdown file in a testdata directory.
These are copied to the target tmp directory and the test
succeeds.

Change-Id: Iae654dca2374b925e6c895a71d4d93b12ccc188b