sorted/postgres: create databases on boot
Previously, `create table ...` sql migrations were being made without the
database existing. This resulted in a panic and error like:

pq: database "pk_3a94488d_blobpacked" does not exist

There seems to be an upstream issue with our postgres library in which
`CREATE DATABASE ...` queries are not prepared so we have to build the
sql manually. For now I've added a regex to make sure we don't allow
anything too crazy in.

Fixes #1022

Change-Id: I0da16759e9219347bb11713b92337021546f9d57