website: properly calculate domainName in prod
move calculation of domainName into main rather than init.  This is
necessary because the root var, which domainName is based on, is not
passed as a command line flag when run in production.

Also clean up a few minor things in godoc.go from previous CLs, such as
removing the docRx var which is no longer used.

Fixes #1036

Change-Id: I7145b87d5f1ed13e143bd3750ce22f2d41d6169d