pkg/index: use the gpg ID (and not its blobRef) in corpus
Otherwise claims that are actually from the same signer, end up being
treated as from different signers, because some of the claims were
signed by the sha1 version and others by the sha224 one.

TODO in follow-up CLs: similar fixes in rest of the corpus, such as with
claimPtrsAttrValue. See if non-corpus index functions/methods suffer
from the same problem.

Change-Id: Icbc70e97edc569f46e575d79aaf4359b33996053