pkg/serverinit: force long GPG identity when generating low-level conf
The index and search pkg now rely on the GPG identity to be in the long
form (16 chars). But the server configuration historically was using the
short form (8 chars).

This change therefore forces the GPG identity ending up in the generated
low-level config to be in the long form.

Also, fix the functions in pkg/jsonsign to generate the long form as well,
to ease the consistent use of it everywhere else.

Fixes #1047

Change-Id: I1333150316caa2c5e179dde0c92b850a362ad17e