pkg/importer/gphotos: use "drive" space as "photos" is deprecated
I have not found any doc stating that the photos Space is not supposed
to work anymore, but Brad said so in the bug, and some testing seemed to
confirm it. So this change switches to using the "drive" Space which
contains all files in Google Drive, including the photos from Google
Photos, as long as the option to have the Google Photos folder in Drive
is enabled.

Fortunately, we can still request the Spaces property of the file, and
use that to filter out all of the non-photos files. For now...

Fixes #987

Change-Id: I7f257b70804fd72c730ded158173ec6ae64fde5b