clients/android: show camput error messages
So far, errors in the uploader configuration leading to upload failures
are not reported to the user in a meaningful way. The upload process
simply stops without further feedback, which is annoying especially when
either certificate hash or password are wrong.

This change adds a text widget below the upload stats, reporting
camput's stderr output as-is. While not ideal from a usability
standpoint (log entries are too wide and too verbose, and not clearly
marked as errors), it is better to receive those verbose messages than
none at all. The error text widget is cleared when a new upload is
started.

Change-Id: I75e79d802cf64981b92c80ce9fc12d2a26552169