pkg/client: check client/server version mismatch
Since the search API changed at
2018-01-13-6e8a5930c9fee81640c6c75a9a549fec98064186, a server built
before that revision is incompatible with a client built at or after
that revision.

This change adds a check that is run when this incompatibility could
manifest by a JSON parsing error, in order to clarify the error message.

As a side note, even if there was not this search API breakage, a server
built before 27bacd3df17fc004bbec69f1374b6aebb2e838c7 would panic anyway
if a client built at or after 38f10a7bd0f97475c2b876a7fde08e4ce47bf224
tried uploading to it. Because of the wrong assumption that br.Hash()
is never nil, which can actually happen after sha224 is introduced
(client-side), and the server does not know about sha224.

Fixes #1074

Change-Id: I3dabf25c0200625e177e0f26308465cd35e8c750