pkg/blobserver/blobpacked: reindex duplicates as well
For reasons yet to be explained, it can happen that zip blobs with the
exact same overall data contents get written to the large blobpacked.
Some of their packed schema blobs might differ though, e.g. if the
file names are different.

This is a problem, at least because of the added storage space used. And
also because it is not supposed to happen as there is a wholeRef check
on reception of a file.

However, it seems harmless to let these duplicates get indexed as z:
rows.

Therefore, this change improves a bit the inspection of duplicates and
the logging around it, to help users delete them by hand if they decide
to do so, but also adds detected duplicates as found zip blobs to the
index. This means finding duplicates is not longer considered a failure
of the recovery process.

Updates issue #1079

Change-Id: I5fec2f2226818eefd03ffce82c768de867fad6b0