internal/images: reset err to nil on no EXIF found
the changes to DecodeConfig in ce955adf485fb2e87df05d260c376c81fec0a7a1
moved the call to image.DecodeConfig to before the call to exif.Decode,
because we need to do that for the HEIF case.

After the call to exif.Decode, if err is nil and is not a short read
error, we ignore the error and the code goes on. To mark that intent,
we should have explicitly set the err to nil at this point. However,
that omission did not matter before
ce955adf485fb2e87df05d260c376c81fec0a7a1 because image.DecodeConfig
was the last setter of the error.

Now that it has moved up, we do have to explicitly reset the err to nil.
While we were at it, we rewrote the whole logic to be easier to read.

This seems to fix all the tests in pkg/index and pkg/search.

Change-Id: Id0b711179ccb1365b851d3e86ecb70d47ac358b4