make.go: handle genfileembed stderr output again
Due to a regression introduced in f8ef262f, the stderr output of
genfileembed was not handled anymore. This caused the build to hang on
Windows systems.

This change introduces a buffer to contain all of genfileembed's stderr
output (usually a long list of files). The buffer's contents are printed
when the command fails to run successfully or when verbose build mode is
enabled. Running of the command is simplified by just calling Run()
instead of Start() and Wait(). The -output-files-stderr flag for
genfileembed, obsolete since f8ef262f, is removed.

Fixes #1099

Change-Id: Ibd8a124c73bf347b92acfc85fffd1186dfe4944b