pkg/index: on slurping bogus gps coordinate, log instead of panicking
If a GPS coordinate is bogus, we should have caught it when receiving
the blob, and not write it out as an index row.
If we did not, we end up catching it when trying to slurp it to the
corpus. At this point there's no reason to penalize the user by
panicking, we can instead skip it and log about the error.

Change-Id: Ie40479befca78d8b5da343ef1a8420f703f07cb7