pkg/test/docker: force the MySQL version to 5
Since we don't support the latest MySQL version, our tests in Travis CI
are failing since they pull the latest docker image.

We should fix the issue
(https://github.com/perkeep/perkeep/issues/1114), but in the meantime,
we might as well restrict the tests to MySQL 5, so that CI tests keep on
being useful to alert us about other errors.

Change-Id: I32a2cc1dc5a4194c891183d96498dee14c759baf