internal/geocode: change Lookup behavior wrt to AltLookupFn
Even when AltLookupFun was defined, Lookup used to fallthrough after the
AltLookupFn, depending on what AltLookupFn returned.
This means in some cases, our tests in pkg/search did end up calling the
Google API (which was not intended). We found out because with the new
dependency on an API key for the Google service, we now rely on reading
the key from the config dir, which is wired to panic if it is accessed
during tests.

This change therefore simplifies Lookup, by making it return the results
of the AltLookupFn call, if AltLookupFn exists.

Change-Id: Ie196f892f36d309c725a6e869094364d49cffdc8