website: revert bin name to camlistored for demo
We have not updated the camlistore/demoblobserver image that the website
runs, so the binary it should execute is camlistored, not perkeepd.
As it is, it keeps on trying to run it every 10s, which makes it busy
for nothing. And I don't think it's worth spending time on updating the
image for now, so it's easier to just revert to the old name.

Change-Id: I769aecaaecdb027d686a78adf9a37a14e9ef9b9d