pkg/serverinit: revert change that enabled https if cert or key was specified
My previous commit (f3f38f0c76) made a change that https became true
if httpsCert or httpsKey was enabled. This broke devcam server, which
always sets the cert+key fields in config/dev-server-config.json but
always set "https" to either true or false.

Instead, make it an error to omit "https" if the cert or key is
specified. (It's still currently legal to specify https==false but
specify a cert or key files, so devcam server now works, but it's not
clear that that's a good idea. I might revisit this later when I
update dev-server-config.json to the upcoming TOML format.)

Fixes #1168

Change-Id: I29d000c5eb9472cc901fbd4cb91159a6c293570e