misc/docker/synology: tag images and build stages
Tagging images depending on the selected arch allows us to preserve
docker caching even when switching arches.

Also, since the first build stage was not tagged, it was considered as
dangling, so it was getting removed by 'docker rmi'. So we now tag it as
well to keep it cached.

Change-Id: I0c36c625e7bd459192f97052d26d76084a03bea7