dev/devcam: use libsqlite3 tag when -sqlite=true
In 5894af5196517e6a8f31bb9d75cfa97e8bbdb153 we chose not to vendor
sqlite3-binding.c and sqlite3-binding.h in, because we rely on the
system's libsqlite3 instead. But we never told devcam test to use the
corresponding necessary "libsqlite3" tag, which is now done in this CL.

Also, comment properly about it in .gitignore.

And make Travis CI use the devcam installed in $GOPATH/bin, instead of
specifically building one in tree.

Change-Id: I4b7bb1305669037f84063c7cf6c7f356c99f3edc