server/perkeepd/ui: treat ref: as special case only with full blobRef
The ref: predicate was treated as a special case, to make it correspond
to navigating to a "blobRef URL", e.g.
http://localhost:3179/ui/sha224-b6a88695e3a9178d45fdfb1fa888939d5adc090f866aaa1d7b5ec3a5

However, we should not do that when the query is only for the prefix of
a blobRef, e.g. "ref:sha224-b6a88695".
This CL therefore adds a check to make sure we're in the presence of a
full blobRef before navigating to the special case.

Fixes #1119

Change-Id: Ib92b43adbb2316c1607d617eb03de132c38de85f