images/haveImageID: use docker images, not inspect
Docker inspect will not exit 0 if no version of the image is available
on the machine, leading to the method failing. There is no flag to
adjust this in docker inspect, the only solution would be to parse the
error message but that would be fragile.

Instead, we opt to use `docker images` which has a very similar output
with the exception that it will list all tags available for that image
and that it will not fail on absent images.