doc: document all environement variables (#1383)
envvardoc: repair the script so that it does not fail on very long
lines.

envvardoc: look for env vars with the `PERKEEP` prefix too.

envvardoc: adapt the `docVar` regex to the markdown format.

envvardoc: also look for vars in internal/.

CI: run envvardoc so that we notice when we add new undocumented
environement variables.

doc: List all remaining undocumented env variables. This "shitlist"
gives us a green build in the CI and allows us to ensure all new
environment variables are documented in the future.

doc: Link to the environment variables page.