web-ui/importer: use relative links for importers and accounts
In some network configurations it's annoying that the links in the importer
web ui point to the configured base URL. If you use VPN like tailcale you
might not have the same URL (like home network IP)to access the server.

This changes the links in the web UI to relatives URLs. Since this method
is only used in the web UI it should be pretty safe.