Introduce pdfcabinet
There was a small amount of discussion about this here:
https://groups.google.com/g/perkeep/c/B_Wq3ovph2I

At this point, it seems to more or less work.  It is pretty much a
ripoff of scanning cabinet, but oriented to work with pdfs rather than
images.  Since a pdf often contains multiple pages and I couldn't
think of a reason for a pdf document to contain multiple pdfs, the
relationship is 1:1.

I thought this is a good place to share what I have and see who it
looks to folks.

I haven't yet implemented the 'who' functionality but don't think that
will be too hard.  Pretty much like tags, but with an attribute
named something like 'pdfcabinet:who'.

Right now when displaying the pdf document, I'm using an <object> tag
to show the associated pdf on the page and it contains all the perkeep
chrome too, which is a bit weird.  If someone can point me at a way to
embed it without the perkeep chrome inside the object, I would
appreciate it.

Another weirdness is that I am currently storing the filename of the
pdf in the pdfcabinet:pdf permanode even though it is also in the file
permanode.  That is because I couldn't figure out how to efficiently
grab it (I need it for each un-annotated pdf so I have something to
identify the pdf to the user on the main page).

I'll probably also want to provide some facility for bulk uploading of
already tagged and dated pdfs, but have not thought hard about this yet.