test: add initial webdav test
Some adaptations had to be made to make perkeepd in `world` be able to
start the webdav app.

* blobstorage handler is expected to be named `bs-recv` for serverinit/devmode.go
* server url needs to be known for the app so that it can fetch its config

To facilitate the second point the start up logic in world.go was
adapted to try starting on a random port, the resulting flakiness should
be acceptable. Obsolete code after this was cleaned up.