index: add receive hook
This removes the need for the `cond` storage and hooks up the index
and its backing storage using the blobhub. Also fixed a bug where
serverinit installs a wrapper instead of the storage which causes
the receive hooks to be fired for the wrapper but not for the storage
itself causing us to mix sync and index for it.