blobserver/files: vendor, use Go's robustio package
Windows renames can be flaky, which can lead to problems when we tried
to add a receive hook for the indexer. To address this we copy the
robustio package from Go's src/cmd/internal/robustio and use it.

The thirdparty/go/robustio package is Go 1.25's robustio, but
updating the internal/syscall/windows package to instead use
golang.org/x/sys/windows.

Signed-off-by: Michael Hoffmann <mhoffm@posteo.de>