Remove CamliNetIP, camnetdns, gpgchallenge
These were all part of an earlier effort to let it be easy for people
to easily deploy their own Perkeep servers and get them on the
internet. The idea was that we'd run a small DNS server which would
map from GPG public keys to their rented cloud IPs which users would
prove to us with the gpgchallenge stuff.

The recently added Tailscale support (see 43f34e5cc511, #1668) makes
most of this redundant.

Also I'd stopped running this infrastructure ages ago and removed the
launcher code recently in b5823a65b9587 (and disabled it in
c9f78d02adf). So this was all basically dead code.

Signed-off-by: Brad Fitzpatrick <brad@danga.com>