Add support for SVG and WEBP. * adds webp format support (one line change) * adds svg support (bulk of the changes) * Fixes frontend bug
The frontend bug is a one-line change, and happens in this scenario:
* the mime magic detects the image as "image/whatever"
* Go image decoder does not support the image format

This causes the indexer to enter a semi-broken state, where its identified the image as an image, but has not assigned it dimensions. If this file is then assigned as a `camliContentImage`, it will break, due to `ccim.image` being undefined.